requirements.txt
streamlit
numpy
pandas
scikit-learn
sentence-transformers
nltk
joblib
groq api
pymupdf
pytesseract
Pillow

